+++ spec/dummy/client/entries/app-kitten.js
@@ -115,7 +115,7 @@ import {
   KarlGridTwelveColumns,
   KarlGridWithOffset,
 } from 'kitten/karl/grid/grid'
-import { KarlLegoGrid } from 'kitten/karl/grid/lego-grid'
+import { KarlLegoGrid, KarlLegoGrid12items } from 'kitten/karl/grid/lego-grid'
 import { KarlRow } from 'kitten/karl/grid/row'
 import { KarlScrollableContainer } from 'kitten/karl/grid/scrollable-container'
 import {
@@ -407,6 +407,7 @@ ReactOnRails.register({
   KarlGridTwelveColumns,
   KarlGridWithOffset,
   KarlLegoGrid,
+  KarlLegoGrid12items,
   KarlRow,
   KarlScrollableContainer,
   KarlSideGridWithLeftAsideAndSmallSize,